+++ src/main/java/com/zeebra/domain/product/entity/FavoriteProduct.java
@@ -1,4 +1,4 @@
-package com.zeebra.domain.reviewLike.entity;
+package com.zeebra.domain.product.entity;
 
 import com.zeebra.global.jpa.BaseEntity;
 import jakarta.persistence.Entity;
@@ -8,22 +8,27 @@ import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+
 @Entity
 @Getter
 @NoArgsConstructor
-public class ReviewLike extends BaseEntity {
+public class FavoriteProduct extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long reviewId;
-
     private Long memberId;
 
-    public ReviewLike(Long id, Long reviewId, Long memberId) {
+    private Long productId;
+
+    private LocalDateTime deletedAt;
+
+    public FavoriteProduct(Long id, Long memberId, Long productId, LocalDateTime deletedAt) {
         this.id = id;
-        this.reviewId = reviewId;
         this.memberId = memberId;
+        this.productId = productId;
+        this.deletedAt = deletedAt;
     }
 }
