+++ src/main/java/com/zeebra/domain/product/entity/OptionName.java
@@ -1,4 +1,4 @@
-package com.zeebra.domain.optionName.entity;
+package com.zeebra.domain.product.entity;
 
 import com.zeebra.global.jpa.BaseEntity;
 import jakarta.persistence.Entity;
