+++ src/main/java/com/zeebra/domain/product/repository/FavoriteProductRepository.java
@@ -1,6 +1,6 @@
-package com.zeebra.domain.favoriteProduct.repository;
+package com.zeebra.domain.product.repository;
 
-import com.zeebra.domain.favoriteProduct.entity.FavoriteProduct;
+import com.zeebra.domain.product.entity.FavoriteProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FavoriteProductRepository extends JpaRepository<FavoriteProduct, Long> {
