+++ src/main/java/com/zeebra/domain/product/repository/OptionCombinationRepository.java
@@ -1,6 +1,6 @@
-package com.zeebra.domain.optionCombination.repository;
+package com.zeebra.domain.product.repository;
 
-import com.zeebra.domain.optionCombination.entity.OptionCombination;
+import com.zeebra.domain.product.entity.OptionCombination;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OptionCombinationRepository extends JpaRepository<OptionCombination, Long> {
