+++ src/main/java/com/zeebra/domain/product/repository/OptionNameRepository.java
@@ -1,6 +1,6 @@
-package com.zeebra.domain.optionName.repository;
+package com.zeebra.domain.product.repository;
 
-import com.zeebra.domain.optionName.entity.OptionName;
+import com.zeebra.domain.product.entity.OptionName;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OptionNameRepository extends JpaRepository<OptionName, Long> {
