+++ src/main/java/com/zeebra/domain/product/repository/ProductOptionRepository.java
@@ -1,6 +1,6 @@
-package com.zeebra.domain.productOption.repository;
+package com.zeebra.domain.product.repository;
 
-import com.zeebra.domain.productOption.entity.ProductOption;
+import com.zeebra.domain.product.entity.ProductOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductOptionRepository extends JpaRepository<ProductOption, Long> {
