+++ src/main/java/com/zeebra/domain/product/repository/SalesRepository.java
@@ -1,6 +1,6 @@
-package com.zeebra.domain.sales.repository;
+package com.zeebra.domain.product.repository;
 
-import com.zeebra.domain.sales.entity.Sales;
+import com.zeebra.domain.product.entity.Sales;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SalesRepository extends JpaRepository<Sales, Long> {
