+++ src/main/java/com/zeebra/domain/review/repository/ReviewLikeRepository.java
@@ -1,6 +1,6 @@
-package com.zeebra.domain.reviewLike.repository;
+package com.zeebra.domain.review.repository;
 
-import com.zeebra.domain.reviewLike.entity.ReviewLike;
+import com.zeebra.domain.review.entity.ReviewLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewLikeRepository extends JpaRepository<ReviewLike, Long> {
